Boons and Banes
-Can now construct Avatar and input Boons and Banes
-Let's the user know if they have selected the same boon and bane when
they try to calculate and does not allow it
-Children can have Avatar as their parent
-Added myBoon and myBane to unitcontroller
-Deleted marriage locked arrrays from DataStorage
-Children math now deals with all doubles in Unit.java
-Deleted unnecessary data from Kana.java

+++ src/domain/Unit.java
@@ -170,7 +170,7 @@ public class Unit {
 			{
 				// FORUMLA: c = Child's absolute base stats + {(n x Child's full growth rates) rounded down}
 				//      Child's absolute base stat							 + ROUNDED DOWN	  (n * full growths while in the character's base class)
-				double c = myChildCharacter.getBaseStats().getStats(route, i+1) + n * ((growths[i] - myJob.getGrowths(i) + tempGrowths[i])/100.0);
+				double c = myChildCharacter.getBaseStats().getStats(route, i+1) + n * ((growths[i] - myJob.getGrowths(i) + (double)tempGrowths[i])/100.0);
 				//System.out.println(statblock[i] + ": " + c);
 				// Calculate the fixedParent's and variedParent's personal stats by subtracting the base stats of their job
 				double fixedParentPersonal = fixedParentStats[i] - fixedParent.getMyJob().getBaseStats(i);
